Promote Decomposition to be public
