+++ modules/math/src/main/java/com/opengamma/strata/math/impl/linearalgebra/DecompositionFactory.java
@@ -8,6 +8,8 @@ package com.opengamma.strata.math.impl.linearalgebra;
 import java.util.HashMap;
 import java.util.Map;
 
+import com.opengamma.strata.math.rootfind.Decomposition;
+
 /**
  * Factory class for different types of decompositions.
  */
