+++ modules/math/src/main/java/com/opengamma/strata/math/impl/linearalgebra/SVDecompositionCommons.java
@@ -11,13 +11,14 @@ import org.apache.commons.math3.linear.SingularValueDecomposition;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleMatrix;
 import com.opengamma.strata.math.impl.util.CommonsMathWrapper;
+import com.opengamma.strata.math.rootfind.Decomposition;
 
 /**
  * This class is a wrapper for the <a href="http://commons.apache.org/math/api-2.1/org/apache/commons/math/linear/SingularValueDecompositionImpl.html">Commons Math library implementation</a>
  * of singular value decomposition.
  */
 // CSOFF: AbbreviationAsWordInName
-public class SVDecompositionCommons extends Decomposition<SVDecompositionResult> {
+public class SVDecompositionCommons implements Decomposition<SVDecompositionResult> {
 
   @Override
   public SVDecompositionResult apply(DoubleMatrix x) {
