+++ modules/math/src/main/java/com/opengamma/strata/math/impl/rootfinding/newton/InverseJacobianEstimateInitializationFunction.java
@@ -10,8 +10,8 @@ import java.util.function.Function;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.collect.array.DoubleMatrix;
-import com.opengamma.strata.math.impl.linearalgebra.Decomposition;
-import com.opengamma.strata.math.impl.linearalgebra.DecompositionResult;
+import com.opengamma.strata.math.rootfind.Decomposition;
+import com.opengamma.strata.math.rootfind.DecompositionResult;
 
 /**
  *
