+++ modules/math/src/test/java/com/opengamma/strata/math/impl/rootfinding/newton/JacobianDirectionFunctionTest.java
@@ -11,8 +11,8 @@ import org.testng.annotations.Test;
 
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.collect.array.DoubleMatrix;
-import com.opengamma.strata.math.impl.linearalgebra.Decomposition;
 import com.opengamma.strata.math.impl.linearalgebra.DecompositionFactory;
+import com.opengamma.strata.math.rootfind.Decomposition;
 
 /**
  * Test.
